backend/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
pymysql
sqlite-utils
requests
python-multipart
python-jose[cryptography]
passlib[bcrypt]
google-generativeai
python-dotenv
cryptography
bcrypt
